error_utils: move error helper to separate header
